component_tests/pytest_appium_android_tests/requirements.txt
pytest==8.3.5
Appium-Python-Client==5.1.2
requests
